combined requestPasswordResetController and resetPasswordController into a single file as the current convention is to have one controller file handle each route file. also removed unnessesary console logs on success in a number of controller functions.
