+++ server/controllers/medicationController.js
@@ -17,7 +17,6 @@ exports.getListOfMedications = function (req, res) {
                     success: false
                 });
             } else {
-                console.log(rows);
                 res.status(200).send(rows);
             }
         });
